# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CTI Demo API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CompanySuggestion {
  name: string;
  domain: string;
  logo?: string | null;
}

export interface DomainCounts {
  total: number;
  employees: number;
  users: number;
  third_parties: number;
  last_employee_compromised?: string | null;
  last_user_compromised?: string | null;
}

export interface UrlEntry {
  url: string;
  type: string;
  occurrence: number;
}

export type DomainUrlsData = {
  employees_urls?: UrlEntry[];
  clients_urls?: UrlEntry[];
  all_urls?: UrlEntry[];
};

export type DomainUrlsStats = {
  totalEmployees?: number;
  totalUsers?: number;
};

export interface DomainUrls {
  data: DomainUrlsData;
  stats: DomainUrlsStats;
  totalUrls: number;
}

export interface AppDetail {
  name: string;
  description: string;
  ai_scenarios?: string[];
}

export type DomainApplicationsApplicationsItem = {
  keyword?: string;
};

export interface DomainApplications {
  applications: DomainApplicationsApplicationsItem[];
  appDetails: AppDetail[];
}

export type PasswordStrengthTooWeak = {
  qty?: number;
  perc?: number;
};

export type PasswordStrengthWeak = {
  qty?: number;
  perc?: number;
};

export type PasswordStrengthMedium = {
  qty?: number;
  perc?: number;
};

export type PasswordStrengthStrong = {
  qty?: number;
  perc?: number;
};

export interface PasswordStrength {
  totalPass: number;
  too_weak?: PasswordStrengthTooWeak;
  weak?: PasswordStrengthWeak;
  medium?: PasswordStrengthMedium;
  strong?: PasswordStrengthStrong;
}

export interface DomainPasswords {
  employeePasswords: PasswordStrength;
  userPasswords: PasswordStrength;
}

export type StealerFamiliesStealerFamilies = { [key: string]: number | string };

export interface StealerFamilies {
  stealerFamilies: StealerFamiliesStealerFamilies;
}

export interface AntivirusEntry {
  count: number;
  name: string;
}

export type AntivirusesAntiviruses = {
  total?: number;
  found?: number;
  not_found?: number;
  free?: number;
  list?: AntivirusEntry[];
};

export interface Antiviruses {
  antiviruses: AntivirusesAntiviruses;
}

export type GetSearchSuggestionsParams = {
  query: string;
};