# lib/api-zod/src/generated/types/antiviruses.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CTI Demo API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AntivirusesAntiviruses } from "./antivirusesAntiviruses";

export interface Antiviruses {
  antiviruses: AntivirusesAntiviruses;
}