# lib/api-zod/src/generated/types/domainApplications.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CTI Demo API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppDetail } from "./appDetail";
import type { DomainApplicationsApplicationsItem } from "./domainApplicationsApplicationsItem";

export interface DomainApplications {
  applications: DomainApplicationsApplicationsItem[];
  appDetails: AppDetail[];
}