# lib/api-zod/src/generated/types/domainCounts.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CTI Demo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DomainCounts {
  total: number;
  employees: number;
  users: number;
  third_parties: number;
  last_employee_compromised?: string | null;
  last_user_compromised?: string | null;
}