# lib/api-zod/src/generated/types/domainUrls.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CTI Demo API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DomainUrlsData } from "./domainUrlsData";
import type { DomainUrlsStats } from "./domainUrlsStats";

export interface DomainUrls {
  data: DomainUrlsData;
  stats: DomainUrlsStats;
  totalUrls: number;
}