# lib/api-zod/src/generated/types/passwordStrength.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CTI Demo API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PasswordStrengthMedium } from "./passwordStrengthMedium";
import type { PasswordStrengthStrong } from "./passwordStrengthStrong";
import type { PasswordStrengthTooWeak } from "./passwordStrengthTooWeak";
import type { PasswordStrengthWeak } from "./passwordStrengthWeak";

export interface PasswordStrength {
  totalPass: number;
  too_weak?: PasswordStrengthTooWeak;
  weak?: PasswordStrengthWeak;
  medium?: PasswordStrengthMedium;
  strong?: PasswordStrengthStrong;
}