# lib/api-zod/src/generated/types/urlEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CTI Demo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UrlEntry {
  url: string;
  type: string;
  occurrence: number;
}